Extend TryCatchScopes to TryCatchFinallyScopes
