+++ gen/funcgenstate.h
@@ -17,7 +17,7 @@
 
 #include "gen/irstate.h"
 #include "gen/pgo.h"
-#include "gen/trycatch.h"
+#include "gen/trycatchfinally.h"
 #include "llvm/ADT/DenseMap.h"
 #include "llvm/IR/CallSite.h"
 #include <vector>
@@ -33,13 +33,6 @@ class Constant;
 class MDNode;
 class Value;
 }
-
-/// Represents a position on the stack of currently active cleanup scopes.
-///
-/// Since we always need to run a contiguous part of the stack (or all) in
-/// order, two cursors (one of which is usually the currently top of the stack)
-/// are enough to identify a sequence of cleanups to run.
-using CleanupCursor = size_t;
 
 /// Stores information needed to correctly jump to a given label or loop/switch
 /// statement (break/continue can be labeled, but are not necessarily).
@@ -89,81 +82,6 @@ struct GotoJump {
            llvm::BasicBlock *tentativeTarget, Identifier *targetLabel);
 };
 
-/// Describes a particular way to leave a cleanup scope and continue execution
-/// with another one.
-///
-/// In general, there can be multiple ones (normal exit, early returns,
-/// breaks/continues, exceptions, and so on).
-struct CleanupExitTarget {
-  explicit CleanupExitTarget(llvm::BasicBlock *t) : branchTarget(t) {}
-
-  /// The target basic block to branch to after running the cleanup.
-  llvm::BasicBlock *branchTarget = nullptr;
-
-  /// The basic blocks that want to continue with this target after running
-  /// the cleanup. We need to keep this information around so we can insert
-  /// stores to the branch selector variable when converting from one to two
-  /// targets.
-  std::vector<llvm::BasicBlock *> sourceBlocks;
-
-  /// MSVC: The basic blocks that are executed when going this route
-  std::vector<llvm::BasicBlock *> cleanupBlocks;
-};
-
-/// Represents a scope (in abstract terms, not curly braces) that requires a
-/// piece of cleanup code to be run whenever it is left, whether as part of
-/// normal control flow or exception unwinding.
-///
-/// This includes finally blocks (which are also generated by the frontend for
-/// running the destructors of non-temporary variables) and the destructors of
-/// temporaries (which are unfortunately not lowered by the frontend).
-///
-/// Our goal is to only emit each cleanup once such as to avoid generating an
-/// exponential number of basic blocks/landing pads for handling all the
-/// different ways of exiting a deeply nested scope (consider e.g. ten
-/// local variables with destructors, each of which might throw itself).
-class CleanupScope {
-public:
-  CleanupScope(llvm::BasicBlock *beginBlock, llvm::BasicBlock *endBlock)
-      : beginBlock(beginBlock), endBlock(endBlock) {}
-
-  /// The basic block to branch to for running the cleanup.
-  llvm::BasicBlock *beginBlock = nullptr;
-
-  /// The basic block that contains the end of the cleanup code (is different
-  /// from beginBlock if the cleanup contains control flow).
-  llvm::BasicBlock *endBlock = nullptr;
-
-  /// The branch selector variable, or null if not created yet.
-  llvm::AllocaInst *branchSelector = nullptr;
-
-  /// Stores all possible targets blocks after running this cleanup, along
-  /// with what predecessors want to continue at that target. The index in
-  /// the vector corresponds to the branch selector value for that target.
-  // Note: This is of course a bad choice of data structure for many targets
-  // complexity-wise. However, situations where this matters should be
-  // exceedingly rare in both hand-written as well as generated code.
-  std::vector<CleanupExitTarget> exitTargets;
-
-  /// Keeps track of all the gotos originating from somewhere inside this
-  /// scope for which we have not found the label yet (because it occurs
-  /// lexically later in the function).
-  // Note: Should also be a dense map from source block to the rest of the
-  // data if we expect many gotos.
-  std::vector<GotoJump> unresolvedGotos;
-
-  /// Caches landing pads generated for catches at this cleanup scope level.
-  ///
-  /// One element is pushed to the back on each time a catch block is entered,
-  /// and popped again once it is left. If the corresponding landing pad has
-  /// not been generated yet (this is done lazily), the pointer is null.
-  std::vector<llvm::BasicBlock *> landingPads;
-
-  /// MSVC: The original basic blocks that are executed for beginBlock to
-  /// endBlock
-  std::vector<llvm::BasicBlock *> cleanupBlocks;
-};
-
 /// Keeps track of active (abstract) scopes in a function that influence code
 /// generation of their contents. This includes cleanups (finally blocks,
 /// destructors), try/catch blocks and labels for goto/break/continue.
@@ -181,10 +99,10 @@ public:
 /// the rest of the ScopeStack API, as it (in contrast to goto) never requires
 /// resolving forward references across cleanup scopes.
 class ScopeStack {
-  friend class TryCatchScopes;
-
 public:
-  explicit ScopeStack(IRState &irs) : irs(irs), tryCatchScopes(irs) {}
+  explicit ScopeStack(IRState &irs) : irs(irs), tryCatchFinallyScopes(irs) {
+    unresolvedGotosPerCleanupScope.emplace_back();
+  }
   ~ScopeStack();
 
   /// Registers a piece of cleanup code to be run.
@@ -192,7 +110,10 @@ public:
   /// The end block is expected not to contain a terminator yet. It will be
   /// added by ScopeStack as needed, based on what follow-up blocks code from
   /// within this scope will branch to.
-  void pushCleanup(llvm::BasicBlock *beginBlock, llvm::BasicBlock *endBlock);
+  void pushCleanup(llvm::BasicBlock *beginBlock, llvm::BasicBlock *endBlock) {
+    tryCatchFinallyScopes.pushCleanup(beginBlock, endBlock);
+    unresolvedGotosPerCleanupScope.emplace_back();
+  }
 
   /// Terminates the current basic block with a branch to the cleanups needed
   /// for leaving the current scope and continuing execution at the target
@@ -200,19 +121,8 @@ public:
   ///
   /// After running them, execution will branch to the given basic block.
   void runCleanups(CleanupCursor targetScope, llvm::BasicBlock *continueWith) {
-    runCleanups(currentCleanupScope(), targetScope, continueWith);
+    tryCatchFinallyScopes.runCleanups(targetScope, continueWith);
   }
-
-  /// Like #runCleanups(), but runs all of them until the top-level scope is
-  /// reached.
-  void runAllCleanups(llvm::BasicBlock *continueWith);
-
-#if LDC_LLVM_VER >= 308
-  void runCleanupCopies(CleanupCursor sourceScope, CleanupCursor targetScope,
-                        llvm::BasicBlock *continueWith);
-  llvm::BasicBlock *runCleanupPad(CleanupCursor scope,
-                                  llvm::BasicBlock *unwindTo);
-#endif
 
   /// Pops all the cleanups between the current scope and the target cursor.
   ///
@@ -224,13 +134,17 @@ public:
   ///
   /// Note that this cursor is only valid as long as the current scope is not
   /// popped.
-  CleanupCursor currentCleanupScope() { return cleanupScopes.size(); }
+  CleanupCursor currentCleanupScope() {
+    return tryCatchFinallyScopes.currentCleanupScope();
+  }
 
   /// Registers a try-catch scope.
-  void pushTryCatch(TryCatchStatement *stmt, llvm::BasicBlock *endbb);
+  void pushTryCatch(TryCatchStatement *stmt, llvm::BasicBlock *endbb) {
+    tryCatchFinallyScopes.pushTryCatch(stmt, endbb);
+  }
 
   /// Unregisters the last registered try-catch scope.
-  void popTryCatch();
+  void popTryCatch() { tryCatchFinallyScopes.popTryCatch(); }
 
   /// Registers a loop statement to be used as a target for break/continue
   /// statements in the current scope.
@@ -292,20 +206,8 @@ public:
   /// the way there.
   void breakToClosest() { jumpToClosest(breakTargets); }
 
-  /// get exisiting or emit new landing pad
-  llvm::BasicBlock *getLandingPad();
-
 private:
-  /// Internal version that allows specifying the scope at which to start
-  /// emitting the cleanups.
-  void runCleanups(CleanupCursor sourceScope, CleanupCursor targetScope,
-                   llvm::BasicBlock *continueWith);
-
   std::vector<GotoJump> &currentUnresolvedGotos();
-
-  std::vector<llvm::BasicBlock *> &currentLandingPads();
-
-  llvm::BasicBlock *&getLandingPadRef(CleanupCursor scope);
 
   /// Unified implementation for labeled break/continue.
   void jumpToStatement(std::vector<JumpTarget> &targets,
@@ -329,23 +231,18 @@ private:
   ///
   std::vector<JumpTarget> continueTargets;
 
-  /// cleanupScopes[i] contains the information to go from
-  /// currentCleanupScope() == i + 1 to currentCleanupScope() == i.
-  std::vector<CleanupScope> cleanupScopes;
-
   ///
-  TryCatchScopes tryCatchScopes;
+  TryCatchFinallyScopes tryCatchFinallyScopes;
 
-  /// Gotos which we were not able to resolve to any cleanup scope, but which
-  /// might still be defined later in the function at top level. If there are
-  /// any left on function exit, it is an error (e.g. because the user tried
-  /// to goto into a finally block, etc.).
-  std::vector<GotoJump> topLevelUnresolvedGotos;
-
-  /// Caches landing pads generated for catches without any cleanups to run
-  /// (null if not yet emitted, one element is pushed to/popped from the back
-  /// on entering/leaving a catch block).
-  std::vector<llvm::BasicBlock *> topLevelLandingPads;
+  /// Keeps track of all the gotos originating from somewhere inside this
+  /// scope for which we have not found the label yet (because it occurs
+  /// lexically later in the function).
+  // Note: Should also be a dense map from source block to the rest of the
+  // data if we expect many gotos.
+  using Gotos = std::vector<GotoJump>;
+  /// The first element represents the stack of unresolved top-level gotos
+  /// (no cleanups).
+  std::vector<Gotos> unresolvedGotosPerCleanupScope;
 };
 
 template <typename T>
@@ -357,7 +254,7 @@ llvm::CallSite ScopeStack::callOrInvoke(llvm::Value *callee, const T &args,
 
   // Ignore 'nothrow' if there are active catch blocks handling non-Exception
   // Throwables.
-  if (isNothrow && tryCatchScopes.isCatchingNonExceptions())
+  if (isNothrow && tryCatchFinallyScopes.isCatchingNonExceptions())
     isNothrow = false;
 
   // Intrinsics don't support invoking and 'nounwind' functions don't need it.
@@ -370,7 +267,7 @@ llvm::CallSite ScopeStack::callOrInvoke(llvm::Value *callee, const T &args,
   llvm::SmallVector<llvm::OperandBundleDef, 2> BundleList;
 #endif
 
-  if (doesNotThrow || (cleanupScopes.empty() && tryCatchScopes.empty())) {
+  if (doesNotThrow || tryCatchFinallyScopes.empty()) {
     llvm::CallInst *call = irs.ir->CreateCall(callee, args,
 #if LDC_LLVM_VER >= 308
                                               BundleList,
@@ -382,7 +279,7 @@ llvm::CallSite ScopeStack::callOrInvoke(llvm::Value *callee, const T &args,
     return call;
   }
 
-  llvm::BasicBlock *landingPad = getLandingPad();
+  llvm::BasicBlock *landingPad = tryCatchFinallyScopes.getLandingPad();
 
   llvm::BasicBlock *postinvoke = irs.insertBB("postinvoke");
   llvm::InvokeInst *invoke =
