+++ gen/trycatchfinally.h
@@ -1,0 +1,227 @@
+//===-- gen/trycatchfinally.h - Try-catch-finally scopes --------*- C++ -*-===//
+//
+//                         LDC – the LLVM D compiler
+//
+// This file is distributed under the BSD-style LDC license. See the LICENSE
+// file for details.
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef LDC_GEN_TRYCATCHFINALLY_H
+#define LDC_GEN_TRYCATCHFINALLY_H
+
+#include <stddef.h>
+#include <vector>
+
+struct IRState;
+class TryCatchStatement;
+
+namespace llvm {
+class AllocaInst;
+class BasicBlock;
+class GlobalVariable;
+class MDNode;
+class Value;
+}
+
+/// Represents a position on the stack of currently active cleanup scopes.
+///
+/// Since we always need to run a contiguous part of the stack (or all) in
+/// order, two cursors (one of which is usually the currently top of the stack)
+/// are enough to identify a sequence of cleanups to run.
+using CleanupCursor = size_t;
+
+////////////////////////////////////////////////////////////////////////////////
+
+class TryCatchFinallyScopes;
+
+class TryCatchScope {
+public:
+  /// Stores information to be able to branch to a catch clause if it matches.
+  ///
+  /// Each catch body is emitted only once, but may be target from many landing
+  /// pads (in case of nested catch or cleanup scopes).
+  struct CatchBlock {
+    /// The ClassInfo reference corresponding to the type to match the
+    /// exception object against.
+    llvm::GlobalVariable *classInfoPtr;
+    /// The block to branch to if the exception type matches.
+    llvm::BasicBlock *bodyBB;
+    // PGO branch weights for the exception type match branch.
+    // (first weight is for match, second is for mismatch)
+    llvm::MDNode *branchWeights;
+  };
+
+  TryCatchScope(TryCatchStatement *stmt, llvm::BasicBlock *endbb,
+                CleanupCursor cleanupScope);
+
+  CleanupCursor getCleanupScope() const { return cleanupScope; }
+  bool isCatchingNonExceptions() const { return catchesNonExceptions; }
+
+  void emitCatchBodies(IRState &irs, TryCatchFinallyScopes &scopes);
+  const std::vector<CatchBlock> &getCatchBlocks() const;
+
+private:
+  TryCatchStatement *stmt;
+  llvm::BasicBlock *endbb;
+  CleanupCursor cleanupScope;
+  bool catchesNonExceptions;
+
+  std::vector<CatchBlock> catchBlocks;
+
+  void emitCatchBodiesMSVC(IRState &irs, TryCatchFinallyScopes &scopes);
+};
+
+////////////////////////////////////////////////////////////////////////////////
+
+/// Represents a scope (in abstract terms, not curly braces) that requires a
+/// piece of cleanup code to be run whenever it is left, whether as part of
+/// normal control flow or exception unwinding.
+///
+/// This includes finally blocks (which are also generated by the frontend for
+/// running the destructors of non-temporary variables) and the destructors of
+/// temporaries (which are unfortunately not lowered by the frontend).
+///
+/// Our goal is to only emit each cleanup once such as to avoid generating an
+/// exponential number of basic blocks/landing pads for handling all the
+/// different ways of exiting a deeply nested scope (consider e.g. ten
+/// local variables with destructors, each of which might throw itself).
+class CleanupScope {
+public:
+  CleanupScope(llvm::BasicBlock *beginBlock, llvm::BasicBlock *endBlock);
+
+  llvm::BasicBlock *run(IRState &irs, llvm::BasicBlock *sourceBlock,
+                        llvm::BasicBlock *continueWith);
+
+  llvm::BasicBlock *runCopying(IRState &irs, llvm::BasicBlock *sourceBlock,
+                               llvm::BasicBlock *continueWith,
+                               llvm::BasicBlock *unwindTo = nullptr,
+                               llvm::Value *funclet = nullptr);
+
+  llvm::BasicBlock *beginBlock() const { return blocks.front(); }
+  llvm::BasicBlock *endBlock() const { return blocks.back(); }
+
+private:
+  std::vector<llvm::BasicBlock *> blocks;
+
+  /// The branch selector variable, or null if not created yet.
+  llvm::AllocaInst *branchSelector = nullptr;
+
+  /// Describes a particular way to leave a cleanup scope and continue execution
+  /// with another one.
+  ///
+  /// In general, there can be multiple ones (normal exit, early returns,
+  /// breaks/continues, exceptions, and so on).
+  struct CleanupExitTarget {
+    explicit CleanupExitTarget(llvm::BasicBlock *t) : branchTarget(t) {}
+
+    /// The target basic block to branch to after running the cleanup.
+    llvm::BasicBlock *branchTarget = nullptr;
+
+    /// The basic blocks that want to continue with this target after running
+    /// the cleanup. We need to keep this information around so we can insert
+    /// stores to the branch selector variable when converting from one to two
+    /// targets.
+    std::vector<llvm::BasicBlock *> sourceBlocks;
+
+    /// MSVC: The basic blocks that are executed when going this route
+    std::vector<llvm::BasicBlock *> cleanupBlocks;
+  };
+
+  /// Stores all possible targets blocks after running this cleanup, along
+  /// with what predecessors want to continue at that target. The index in
+  /// the vector corresponds to the branch selector value for that target.
+  // Note: This is of course a bad choice of data structure for many targets
+  // complexity-wise. However, situations where this matters should be
+  // exceedingly rare in both hand-written as well as generated code.
+  std::vector<CleanupExitTarget> exitTargets;
+};
+
+////////////////////////////////////////////////////////////////////////////////
+
+class TryCatchFinallyScopes {
+public:
+  TryCatchFinallyScopes(IRState &irs);
+
+  bool empty() const { return tryCatchScopes.empty() && cleanupScopes.empty(); }
+
+  void pushTryCatch(TryCatchStatement *stmt, llvm::BasicBlock *endbb);
+  void popTryCatch();
+
+  /// Indicates whether there are any active catch blocks that handle
+  /// non-Exception Throwables.
+  bool isCatchingNonExceptions() const;
+
+  /// Registers a piece of cleanup code to be run.
+  ///
+  /// The end block is expected not to contain a terminator yet. It will be
+  /// added by ScopeStack as needed, based on what follow-up blocks code from
+  /// within this scope will branch to.
+  void pushCleanup(llvm::BasicBlock *beginBlock, llvm::BasicBlock *endBlock);
+
+  /// Terminates the current basic block with a branch to the cleanups needed
+  /// for leaving the current scope and continuing execution at the target
+  /// scope stack level.
+  ///
+  /// After running them, execution will branch to the given basic block.
+  void runCleanups(CleanupCursor targetScope, llvm::BasicBlock *continueWith) {
+    runCleanups(currentCleanupScope(), targetScope, continueWith);
+  }
+
+  /// Runs the single cleanup at the specified stack level and returns the
+  /// beginning basic block of the cleanup code.
+  llvm::BasicBlock *runCleanup(CleanupCursor targetScope,
+                               llvm::BasicBlock *sourceBlock,
+                               llvm::BasicBlock *continueWith);
+
+  /// Pops all the cleanups between the current scope and the target cursor.
+  ///
+  /// This does not insert any cleanup calls, use #runCleanups() beforehand.
+  void popCleanups(CleanupCursor targetScope);
+
+  /// Returns a cursor that identifies the current cleanup scope, to be later
+  /// used with #runCleanups() et al.
+  ///
+  /// Note that this cursor is only valid as long as the current scope is not
+  /// popped.
+  CleanupCursor currentCleanupScope() const { return cleanupScopes.size(); }
+
+  /// Gets the landing pad for the current catches and cleanups.
+  /// If there's no cached one, a new one will be emitted.
+  llvm::BasicBlock *getLandingPad();
+
+private:
+  IRState &irs;
+  std::vector<TryCatchScope> tryCatchScopes;
+
+  /// cleanupScopes[i] contains the information to go from
+  /// currentCleanupScope() == i + 1 to currentCleanupScope() == i.
+  std::vector<CleanupScope> cleanupScopes;
+
+  using LandingPads = std::vector<llvm::BasicBlock *>;
+  /// Landing pads are cached via a dedicated stack for each cleanup scope (one
+  /// element is pushed to/popped from the back on entering/leaving a try-catch
+  /// block).
+  /// The first element represents the stack of top-level landing pads (no
+  /// cleanups).
+  std::vector<LandingPads> landingPadsPerCleanupScope;
+
+  llvm::BasicBlock *&getLandingPadRef(CleanupCursor scope);
+
+  /// Emits a landing pad to honor all the active cleanups and catches.
+  llvm::BasicBlock *emitLandingPad();
+
+  /// Internal version that allows specifying the scope at which to start
+  /// emitting the cleanups.
+  void runCleanups(CleanupCursor sourceScope, CleanupCursor targetScope,
+                   llvm::BasicBlock *continueWith);
+
+  // MSVC
+  llvm::BasicBlock *emitLandingPadMSVC(CleanupCursor cleanupScope);
+  void runCleanupCopies(CleanupCursor sourceScope, CleanupCursor targetScope,
+                        llvm::BasicBlock *continueWith);
+  llvm::BasicBlock *runCleanupPad(CleanupCursor scope,
+                                  llvm::BasicBlock *unwindTo);
+};
+
+#endif
